Move `SuperInterface` out of `settings` package
