+++ mc-java-base/src/main/proto/spine/tools/mc/java/settings.proto
@@ -231,24 +231,6 @@ message Uuids {
     protodata.Actions actions = 3;
 }
 
-// Specification of an existing super interface a generated Java type should implement.
-message SuperInterface {
-
-    // The name of the interface.
-    //
-    // It could be a simple name if the generated types are in the same Java package.
-    // Otherwise, this value must contain a fully qualified canonical name.
-    //
-    string name = 1 [(required) = true];
-
-    // Optional arguments for generic parameters, if the interface has them.
-    //
-    // As for the interface name, the arguments could be simple names if corresponding
-    // types reside in the same package with the generated code.
-    // Otherwise, fully qualified canonical names must be provided.
-    //
-    repeated string generic_argument = 2;
-}
 
 // Configuration for adding a Java interface to a generated class.
 message AddInterface {
