+++ mc-java-marker/src/main/kotlin/io/spine/tools/mc/java/marker/EveryIsMessagesRenderer.kt
@@ -33,10 +33,11 @@ import io.spine.protodata.java.JavaRenderer
 import io.spine.protodata.java.file.hasJavaRoot
 import io.spine.protodata.java.qualifiedJavaType
 import io.spine.protodata.renderer.SourceFileSet
+import io.spine.tools.java.reference
 import io.spine.tools.mc.java.CreateInterface
 import io.spine.tools.mc.java.ImplementInterface
-import io.spine.tools.mc.java.settings.SuperInterface
-import io.spine.tools.mc.java.settings.superInterface
+import io.spine.tools.mc.java.SuperInterface
+import io.spine.tools.mc.java.superInterface
 import io.spine.tools.psi.java.execute
 import org.checkerframework.checker.signature.qual.FullyQualifiedName
 
@@ -45,7 +46,9 @@ internal class EveryIsMessagesRenderer : JavaRenderer() {
     /**
      * The super base interface for all generated interfaces.
      */
-    private val superBase = ClassName(Message::class.java)
+    private val superBase = superInterface {
+        name = Message::class.java.reference
+    }
 
     private lateinit var sources: SourceFileSet
